+++ lib/procon_bypass_man/procon.rb
@@ -50,8 +50,22 @@ class ProconBypassMan::Procon
     analog_stick = ProconBypassMan::Procon::AnalogStick.new(binary: user_operation.binary.raw)
     dumped_tilting_power = @@left_stick_tilting_power_scaler.add_sample(analog_stick.relative_hypotenuse)
 
-    if ongoing_macro.finished?
+    enable_all_macro = true
+    enable_macro_map = Hash.new {|h,k| h[k] = true }
+    current_layer.disable_macros.each do |disable_macro|
+      if (disable_macro[:if_pressed] == [true] || user_operation.pressing_all_buttons?(disable_macro[:if_pressed]))
+        if disable_macro[:name] == :all
+          enable_all_macro = false
+        else
+          enable_macro_map[disable_macro[:name]] = false
+        end
+      end
+    end
+
+    if ongoing_macro.finished? && enable_all_macro
       current_layer.macros.each do |macro_name, options|
+        next unless enable_macro_map[macro_name]
+
         if options[:if_tilted_left_stick]
           if dumped_tilting_power&.tilting?(current_position_x: analog_stick.relative_x, current_position_y: analog_stick.relative_y) && user_operation.pressing_all_buttons?(options[:if_pressed])
             @@status[:ongoing_macro] = MacroRegistry.load(macro_name)