+++ spec/lib/procon_bypass_man/procon/macro_builder_spec.rb
@@ -41,14 +41,14 @@ describe ProconBypassMan::Procon::MacroBuilder do
       describe 'toggle_x_for_2sec' do
         it do
           expect(described_class.new([:toggle_r_for_2sec]).build).to eq(
-            [{ continue_for: 2, steps: [:none, :r] }]
+            [{ continue_for: 2, steps: [:r, :none] }]
           )
         end
         it do
           expect(described_class.new([:toggle_r, :toggle_r_for_3sec]).build).to eq(
             [ :r,
               :none,
-              { continue_for: 3, steps: [:none, :r] }
+              { continue_for: 3, steps: [:r, :none] }
             ]
           )
         end
@@ -57,14 +57,14 @@ describe ProconBypassMan::Procon::MacroBuilder do
       describe 'toggle_x_for_0_2sec' do
         it do
           expect(described_class.new([:toggle_r_for_0_2sec]).build).to eq(
-            [{ continue_for: 0.2, steps: [:none,:r] }]
+            [{ continue_for: 0.2, steps: [:r, :none] }]
           )
         end
         it do
           expect(described_class.new([:toggle_r, :toggle_r_for_0_3sec]).build).to eq(
             [ :r,
               :none,
-              { continue_for: 0.3, steps: [:none, :r] }
+              { continue_for: 0.3, steps: [:r, :none] }
             ]
           )
         end
@@ -98,8 +98,8 @@ describe ProconBypassMan::Procon::MacroBuilder do
         it do
           expect(described_class.new([:pressing_r_and_toggle_zr]).build).to eq(
             [{ continue_for: nil, steps: [
-              [:r, :none],
               [:r, :zr],
+              [:r, :none],
             ] }]
           )
         end
@@ -109,8 +109,8 @@ describe ProconBypassMan::Procon::MacroBuilder do
         it do
           expect(described_class.new([:pressing_x_and_toggle_zr_for_0_2sec]).build).to eq(
             [{ continue_for: 0.2, steps: [
-              [:x, :none],
               [:x, :zr],
+              [:x, :none],
             ] }]
           )
         end
@@ -120,8 +120,8 @@ describe ProconBypassMan::Procon::MacroBuilder do
         it do
           expect(described_class.new([:pressing_x_and_toggle_zr_for_0_2sec]).build).to eq(
             [{ continue_for: 0.2, steps: [
-              [:x, :none],
               [:x, :zr],
+              [:x, :none],
             ] }]
           )
         end
@@ -131,8 +131,8 @@ describe ProconBypassMan::Procon::MacroBuilder do
         it do
           expect(described_class.new([:toggle_x_and_toggle_zr_for_0_2sec]).build).to eq(
             [{ continue_for: 0.2, steps: [
-              [:none, :none],
               [:x, :zr],
+              [:none, :none],
             ] }]
           )
         end
@@ -153,8 +153,8 @@ describe ProconBypassMan::Procon::MacroBuilder do
         it do
           expect(described_class.new([:pressing_thumbr_and_toggle_zr_for_0_6sec]).build).to eq(
             [{ continue_for: 0.6, steps: [
-              [:thumbr, :none],
               [:thumbr, :zr],
+              [:thumbr, :none],
             ] }]
           )
         end
@@ -164,8 +164,8 @@ describe ProconBypassMan::Procon::MacroBuilder do
         it do
           expect(described_class.new([:pressing_zr_and_toggle_b_for_0_6sec]).build).to eq(
             [{ continue_for: 0.6, steps: [
-              [:zr, :none],
               [:zr, :b],
+              [:zr, :none],
             ] }]
           )
         end
@@ -175,8 +175,8 @@ describe ProconBypassMan::Procon::MacroBuilder do
         it do
           expect(described_class.new([:pressing_zr_and_toggle_b_for_0_65sec]).build).to eq(
             [{ continue_for: 0.65, steps: [
-              [:zr, :none],
               [:zr, :b],
+              [:zr, :none],
             ] }]
           )
         end
@@ -184,8 +184,8 @@ describe ProconBypassMan::Procon::MacroBuilder do
         it do
           expect(described_class.new([:pressing_zr_and_toggle_b_for_0_65]).build).to eq(
             [{ continue_for: 0.65, steps: [
-              [:zr, :none],
               [:zr, :b],
+              [:zr, :none],
             ] }]
           )
         end